backend/requirements.txt
fastapi
uvicorn
gitpython
pydantic
pydantic-settings
python-dotenv
google-auth
requests
